+++ app/features/devices/__init__.py
@@ -4,8 +4,9 @@ from .login_service import DeviceLoginService, normalize_device_type
 from .classification import device_type_for_role, normalize_device_role
 from .repository import DeviceRepository
 from .service import DeviceService
+from .save_config_service import SaveConfigService
 
 __all__ = [
     "DeviceLoginService", "DeviceRepository", "DeviceService", "normalize_device_type",
-    "device_type_for_role", "normalize_device_role",
+    "device_type_for_role", "normalize_device_role", "SaveConfigService",
 ]